requirements.txt
# Default
python-dotenv
sentence_transformers
streamlit
llama-index
qdrant_client
llama-index-embeddings-huggingface
fastembed

-e